+++ support-frontend/assets/pages/[countryGroupId]/student/StudentLandingPage.tsx
@@ -6,8 +6,10 @@ import { Header } from 'components/headers/simpleHeader/simpleHeader';
 import { PageScaffold } from 'components/page/pageScaffold';
 import type { GeoId } from 'pages/geoIdConfig';
 import { StudentTsAndCs } from './components/studentTsAndCs';
+import { getStudentTsAndCs } from './helpers/studentTsAndCsCopy';
 
 export function StudentLandingPage({ geoId }: { geoId: GeoId }) {
+	const tsAndCsItems = getStudentTsAndCs(geoId);
 	return (
 		<PageScaffold
 			header={<Header />}
@@ -17,7 +19,7 @@ export function StudentLandingPage({ geoId }: { geoId: GeoId }) {
 				</FooterWithContents>
 			}
 		>
-			<StudentTsAndCs geoId={geoId} />
+			{tsAndCsItems && <StudentTsAndCs tsAndCsItem={tsAndCsItems} />}
 		</PageScaffold>
 	);
 }
